+++ jwst/regtest/conftest.py
@@ -2,16 +2,13 @@ from datetime import datetime
 import os
 import copy
 import json
-import pprint
+import sys
 
 import getpass
 import pytest
-from ci_watson.artifactory_helpers import (
-    get_bigdata_root,
-    get_bigdata,
-    BigdataError,
-    UPLOAD_SCHEMA,
-)
+from ci_watson.artifactory_helpers import UPLOAD_SCHEMA
+
+from .regtestdata import RegtestData
 
 
 TODAYS_DATE = datetime.now().strftime("%Y-%m-%d")
@@ -19,6 +16,7 @@ TODAYS_DATE = datetime.now().strftime("%Y-%m-%d")
 
 @pytest.fixture(scope="session")
 def artifactory_repos(pytestconfig):
+    """Provides Artifactory inputs_root and results_root"""
     inputs_root = pytestconfig.getini('inputs_root')[0]
     results_root = pytestconfig.getini('results_root')[0]
     return inputs_root, results_root
@@ -43,7 +41,7 @@ def postmortem(request, fixturename):
     """
     if request.node.report_setup.passed:
         if request.node.report_call.failed:
-            return request.node.funcargs.get(fixturename)
+            return request.node.funcargs.get(fixturename, None)
 
 
 @pytest.fixture(scope='function', autouse=True)
@@ -73,12 +71,16 @@ def generate_artifactory_json(request, artifactory_repos):
 
     rtdata = postmortem(request, 'rtdata') or postmortem(request, 'rtdata_module')
     if rtdata:
-        path = rtdata.output
-        cwd, _ = os.path.split(path)
-        remote_results_path = artifactory_result_path()
+        cwd = os.path.dirname(rtdata.output)
+        rtdata.remote_results_path = artifactory_result_path()
+        rtdata.test_name = request.node.name
+        # Dump the failed test traceback into rtdata
+        rtdata.traceback = str(request.node.report_call.longrepr)
 
-        upload_schema_pattern.append(path)
-        upload_schema = generate_upload_schema(upload_schema_pattern, remote_results_path)
+        upload_schema_pattern.append(rtdata.input)
+        upload_schema_pattern.append(rtdata.output)
+        upload_schema = generate_upload_schema(upload_schema_pattern,
+            rtdata.remote_results_path)
 
         # Write the upload schema to JSON file
         jsonfile = os.path.join(cwd, "{}_results.json".format(request.node.name))
@@ -86,13 +88,19 @@ def generate_artifactory_json(request, artifactory_repos):
             json.dump(upload_schema, outfile, indent=2)
 
 
-        okify_schema_pattern.append(rtdata.truth)
+        pattern = os.path.join(rtdata.remote_results_path, os.path.basename(rtdata.output))
+        okify_schema_pattern.append(pattern)
         okify_schema = generate_upload_schema(okify_schema_pattern, rtdata.truth_remote)
 
         # Write the okify schema to JSON file
         jsonfile = os.path.join(cwd, "{}_okify.json".format(request.node.name))
         with open(jsonfile, 'w') as outfile:
             json.dump(okify_schema, outfile, indent=2)
+
+        # Write the rtdata class out as an ASDF file
+        path = os.path.join(cwd, "{}_rtdata.asdf".format(request.node.name))
+        rtdata.to_asdf(path)
+        print(rtdata, file=sys.stderr)
 
 
 def generate_upload_schema(pattern, target, recursive=False):
@@ -151,132 +159,13 @@ def jail(request, tmpdir_factory):
     temporary directory, and then have the tests access them.
     """
     old_dir = os.getcwd()
-    newpath = tmpdir_factory.mktemp(request.module.__name__.split('.')[-1] +
-        "_" + request._parent_request.fixturename)
+    path = request.module.__name__.split('.')[-1]
+    if request._parent_request.fixturename is not None:
+        path = path + "_" + request._parent_request.fixturename
+    newpath = tmpdir_factory.mktemp(path)
     os.chdir(str(newpath))
     yield newpath
     os.chdir(old_dir)
-
-
-class RegtestData:
-    """Defines data paths on Artifactory and data retrieval methods"""
-
-    def __init__(self, env="dev", inputs_root="jwst-pipeline",
-        results_root="jwst-pipeline-results", docopy=True):
-        self._env = env
-        self._inputs_root = inputs_root
-        self._results_root = results_root
-
-        self.docopy = docopy
-
-        self._input_remote = None
-        self._truth_remote = None
-        self._input = None
-        self._truth = None
-        self._output = None
-        self._bigdata_root = get_bigdata_root()
-
-    def __repr__(self):
-        return pprint.pformat(
-            dict(input=self.input, output=self.output, truth=self.truth,
-            input_remote=self.input_remote, truth_remote=self.truth_remote),
-            indent=2
-        )
-
-    @property
-    def input_remote(self):
-        if self._input_remote is not None:
-            return os.path.join(*self._input_remote)
-        else:
-            return None
-
-    @input_remote.setter
-    def input_remote(self, value):
-        self._input_remote = value.split(os.sep)
-
-    @property
-    def truth_remote(self):
-        if self._truth_remote is not None:
-            return os.path.join(*self._truth_remote)
-        else:
-            return None
-
-    @truth_remote.setter
-    def truth_remote(self, value):
-        self._truth_remote = value.split(os.sep)
-
-    @property
-    def input(self):
-        return self._input
-
-    @input.setter
-    def input(self, value):
-        self._input = os.path.abspath(value)
-
-    @property
-    def truth(self):
-        return self._truth
-
-    @truth.setter
-    def truth(self, value):
-        self._truth = os.path.abspath(value)
-
-    @property
-    def output(self):
-        return self._output
-
-    @output.setter
-    def output(self, value):
-        self._output = os.path.abspath(value)
-
-    @property
-    def bigdata_root(self):
-        return self._bigdata_root
-
-    @bigdata_root.setter
-    def bigdata_root(self, value):
-        return NotImplementedError("Set TEST_BIGDATA environment variable "
-            "to change this value.")
-
-    # The methods
-    def get_data(self, path=None):
-        """Copy data from Artifactory remote resource to the CWD
-
-        Updates self.input and self.input_remote upon completion
-        """
-        if path is None:
-            path = self.input_remote
-        else:
-            self.input_remote = path
-        self.input = get_bigdata(self._inputs_root, self._env,
-            os.path.dirname(path), os.path.basename(path), docopy=self.docopy)
-
-        return self.input
-
-    def get_truth(self, path=None):
-        """Copy truth data from Artifactory remote resource to the CWD/truth
-
-        Updates self.truth and self.truth_remote on completion
-        """
-        if path is None:
-            path = self.truth_remote
-        else:
-            self.truth_remote = path
-        os.makedirs('truth', exist_ok=True)
-        os.chdir('truth')
-        try:
-            self.truth = get_bigdata(self._inputs_root, self._env,
-                os.path.dirname(path), os.path.basename(path), docopy=self.docopy)
-            self.truth_remote = os.path.join(self._bigdata_root, self._inputs_root, self._env, path)
-        except BigdataError:
-            os.chdir('..')
-            raise
-        os.chdir('..')
-
-        return self.truth
-
-    def get_association(self, asn):
-        raise NotImplementedError()
 
 
 def _rtdata_fixture_implementation(artifactory_repos, envopt, request):