+++ CHANGES.rst
@@ -6,6 +6,11 @@ associations
 
 - Update act_id format to allow base 36 values in product name [#4282]
 
+datamodels
+----------
+
+- Force data model type setting on save [#4318]
+
 master_background
 -----------------
 
@@ -62,6 +67,11 @@ datamodels
 ----------
 
 - Updated the list of allowed NIRCam CORONMSK values in model schemas. [#4234]
+
+flat_field
+----------
+ - Updated handling of error arrays for FGS Guider data, which has not been run
+   through ramp fitting [#4309]
 
 lib
 ---